Seperates header into cpp and headers
